+++ actix-http/CHANGES.md
@@ -1,5 +1,18 @@
 # Changes
 
+## [0.2.0] - 2019-05-12
+
+### Changed
+
+* Update actix-service to 0.4
+
+* Expect and upgrade services accept `ServerConfig` config.
+
+### Deleted
+
+* `OneRequest` service
+
+
 ## [0.1.5] - 2019-05-04
 
 ### Fixed